refactor: add size logs as a feature
